Add func to run the program

+++ chip8.h
@@ -17,7 +17,7 @@ class Chip8 {
         uint8_t memory_[4096];
         uint16_t index_register{0};
         uint16_t program_counter{0};
-        uint8_t stack_[16]{};
+        uint16_t stack_[16]{};
         uint8_t stack_pointer{0};
         uint8_t delay_timer{0};
         uint8_t sound_timer{};
